fix: address code review findings across security, correctness, and maintainability
- Fix path traversal bypass via iterative URI decoding in StaticAssetSchemeHandler
- Fix stream leak in InfiniFrameHttpHandler by using async and proper cancellation
- Fix DisposeAsync race condition with try/finally guard in InfiniFrameWindow
- Fix validator side effects: move directory creation out of FluentValidation rule
- Fix fire-and-forget exception swallowing in RegisterWindowCreatedUtility
- Fix integer overflow in MonitorsUtility area calculation
- Add thread safety to InfiniFrameUriSecurityPolicyRegistry.ConfigureForBuilder
- Use DisposableCompositeFileProvider to dispose PhysicalFileProvider on shutdown
- Add InfiniFrameNativeInteropException for native interop errors (replaces ApplicationException)
- Add XML escaping in TempTargetsFile to prevent MSBuild injection
- Add missing MIME types (.wasm, .mp4, .webm, .webp, .avif)
- Align ExceptionsUtility between Shared and Tools.Pack
- Add LastAddComponentException to InfiniFrameJsComponentConfiguration for error visibility
- Fix InfiniNotificationOperation catch block for new exception type

+++ src/InfiniFrame.Shared/Utilities/MonitorsUtility.cs
@@ -49,7 +49,7 @@ internal static class MonitorsUtility {
         monitor = default;
         if (monitors.IsDefaultOrEmpty) return false;
 
-        long windowArea = Math.Max(0, windowBounds.Width);
+        long windowArea = Math.Max(0, (long)windowBounds.Width);
         windowArea *= Math.Max(0, windowBounds.Height);
 
         int bestIndex = -1;
@@ -62,7 +62,7 @@ internal static class MonitorsUtility {
             Rectangle intersection = Rectangle.Intersect(m.MonitorArea, windowBounds);
             long overlap = 0;
             if (intersection.Width > 0 && intersection.Height > 0) {
-                overlap = (long)intersection.Width * intersection.Height;
+                overlap = (long)intersection.Width * (long)intersection.Height;
             }
 
             // fraction of the *window* that lies on this monitor
